Move getBlocksInTopologicalSorting function to PipelineWrapper
